Move S3PutObjectRequest to put_object mod
